# 테스트를 위한 유닛 더미 데이터 추가
## DataTableManager
- 데이터 테이블 파싱
- 파싱한 데이터 관리

## UnitOnHunt, Monster
- Reset할때 캐릭터 외형 프리펩 교체

## UnitManager
- 가챠 기능 추가 중

+++ Assets/Scripts/Unit/Character/UnitStatsData.cs
@@ -6,7 +6,7 @@ public class UnitStatsData
 {
     [field: SerializeField] public string Name { get; set; }
     [field: SerializeField] public int Id { get; set; }
-    [field: SerializeField] public float GachaChance { get; set; }
+    [field: SerializeField] public int GachaChance { get; set; }
     [field: SerializeField] public int GachaLevel { get; set; }
     [field: SerializeField] public UNIT_JOB Job { get; set; }
     [field: SerializeField] public ATTACK_TYPE BasicAttackType { get; set; }
